+++ docs/html/search/classes_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['ieventhandler_85',['IEventHandler',['../classace__button_1_1IEventHandler.html',1,'ace_button']]]
+  ['ieventhandler_91',['IEventHandler',['../classace__button_1_1IEventHandler.html',1,'ace_button']]]
 ];